For 204 response the body should be empty, not an empty ByteArrayInputStream.
When proxying through the server a body here will result in a
`Transfer-Encoding:chunked` in the aleph server, basically stating the body size
is unknown. Some caching products, like varnish, fails miserably here, e.g. gives
the calling client back an http 503.

With this change, at least when the body is empty it is removed, and the
`Transfer-Encoding:chunked` is not added.