Attach idle handlers before other handlers
Having the idle handlers after the other handlers results in them
timing out the connection after the timeout without any consideration
of actual traffic. This can lead to a socket being closed while it is
still actively being used. Attaching the idle handlers first means
that a connection will only be treated as idle if no traffic has been
sent through the socket for a given period of time.