Remove test dependency on /usr/share/dict/words
This file is not present by default on many systems and in fact, on
systems like NixOS, it's not trivial to install at that path. Also,
there is an unchecked reliance on that file to be at least 1e7 bytes
long. If it is in fact shorter, the tests relying on it will just
silently run with shorter inputs.

Instead, just randomly generate printable ASCII characters and
newlines. This seems to be of comparable performance and doesn't have
the above problems.