Fix assertions in aleph.tcp-ssl-test
Some assertions were ineffective since they would be executed on a
different thread from the test runner. For example, deliberately
introducing a failing assertion wouldn't make the test fail. The bug
can also be observed in the test summary. When running
`test-ssl-echo`, it mentions only one assertion:

> Ran 1 assertions, in 1 test functions

With the patch applied this becomes:

> Ran 3 assertions, in 1 test functions