Improve TCP+SSL server connection handling
When starting an TCP+SSL server, the handler would be called when the
underlying TCP connection is established but before the SSL handshake
has taken place.  At that point, the client's `:ssl-session` value is
still a placeholder (e.g. `.isValid` returns `false`, the cipher suite
is reported as `"SSL_NULL_WITH_NULL_NULL"` and the peer principal is
not available). Also, the stream is not ready for use, so any writes
would be buffered until after the handshake is complete. Since there
is no way to be notified of the handshake completion, the existing
test case worked around this by looking up the SSL session only after
having received a message for the first time which implies that the
session has been successfully established since. However, this is a
pretty awkward API and only works for protocols in which the client
actually is the first to transmit anything. There's also a bit of a
gotcha: The value of `:ssl-session` mutably changes to a different
value after the handshake, so holding on to the initial placeholder
value and only starting to use it later won't do the trick.

This patch changes the semantics so that the handler is only invoked
after the session was successfully established. While technically a
breaking change, it appears unlikely that any existing users would
rely on the original behavior since the placeholder SSL session value
is pretty much useless. The new behavior should be more in line with
user expectation and arguably is more correct (e.g. the handler
doesn't have to wait for I/O before being able to peruse the SSL
session info).