Improve short-circuiting behavior of `wrap-future`
If the Netty future passed to `wrap-future` is already done, resolve
the deferred result immediately. This short-circuits the Netty future
listener indirection and all the overhead it involves in such a
situation.

Note that this optimization was already partially implemented, namely
for the success case. This patch merely generelizes it to all "done"
cases.

Fixes #614.