Add minimal tools.deps support
Just enough to allow using aleph as a git dependency via
tools.deps. Both the deps.edn file and build.clj's inputs are
generated from project.clj which remains the source of truth for
now. Can be used as a starting point for a potential full switch to
tools.deps in the future.

Also included is a script for re-generating these files, as well as a
CI check which should ensure that the generated files don't
diverge. Note that these are intentionally not written as tools.deps
or babashka tasks so as not to introduce additional dev and build
dependencies.