Pin babashka version
This was prompted by a failing CI job after merging
https://github.com/clj-commons/aleph/pull/635 Here, babashka's master
branch had changed between the last CI run on the PR and the run after
merge. Apparently this included some change in the pretty-printer's
behavior, resulting in syntactically different `deps.edn` files. This
illustrates why having such moving dependencies isn't great for CI, so
we pin it to the current master commit now.

Also included is the resulting changes for the generated `deps.edn`.