Enable TLS endpoint identification by default (breaking)
TLS clients now check that the hostname they connect to matches the identity given in the server's
certificate. This prevents man-in-the-middle attacks in cases where the server presents a
cryptographically valid certificate for a different hostname.

All TLS clients use the "HTTPS" algorithm as specified in RFC 2818, section 3.1 by default. This is
a reasonable default even for non-HTTPS cases but can be overridden via the new
`:ssl-endpoint-id-alg` client option.

This is technically a breaking change since TLS clients will now refuse to connect to misconfigured
servers. If desired, the old behavior can be restored by passing `:ssl-endpoint-id-alg nil` to
disable endpoint identification.

Mitigates CVE-2023-4586
Closes #688