Update request timeout test
Moves the sleep from the body to before the return of the Ring map.

Unlike the HTTP1 code, the H2 code can respond as soon as it gets a
header, which prevents the request timeout from triggering.

Also adds a 1.5s delay to server shutdown, so repeating tasks like
(date-header-value) can finish and not throw exceptions.