Fix idle timeout error
Update the client :channel-inactive method in H2 to set the response-d
to closed, in order to mimic HTTP1 behavior

In HTTP1 code, when the client channel is closed, it closes the responses
stream, which leads the take! to get ::closed.

Also replace ::closed with :aleph/closed, so it can be used in other nses