Fix bug in HTTP2 handling of client-pool
The HTTP2 code didn't use the :on-closed callback like the HTTP1 code
did. As a result, closed connections weren't automatically being
disposed of in the connection pool, but reused, which leads to immediate
failures.

The HTTP1 code attached the on-closed cb to the responses stream, but H2
doesn't need a stream. The code now attaches the close cb to the Netty
Channel's close Future. This should be safe, tho there may be some churn
in Dirigiste.

(Also realiased confusing aleph.ssl in tests to test-ssl)