Add HTTP/2 compression support
H2 compression is not well-supported by Netty for its multiplex setup,
only its older single-pipeline H2 classes.

It happens in two parts: (1) decorators are attached to the
Http2FrameCodec that handle the actual de/compression, and (2) a handler
that parses and updates the *-encoding headers is added to each stream
pipeline.

Adds new `compression-options` for both H1/H2, and deprecates
`compression-level` to H1-only, like Netty.