Set up ALPN configuration based on http-versions when possible
This also adds an `http-versions` option to `aleph.http/start-server`. Its value will be used to
automatically set up a matching ALPN config when `ssl-context` is a map and doesn't contain a
`:application-protocol-config` key, yet. Support for the latter is retained so that users can still
override the default in case any more interesting settings will become available in the future.

If an `:application-protocol-config` is present, we assert that the configured protocols match the
desired HTTP versions. We need to have that validation code path anyway for the case when an
`io.netty.handler.ssl.SslContext` instance is passed, which made the decision to retain the
`:application-protocol-config` option easier.