Centralize :port and :shutdown-timeout settings in http-test
Forgetting to pass `:shutdown-timeout` to a test server (as was the case in
`test-ssl-session-access` for example) adds 15s extra wait time for each instance. This patch
centralizes setting it to 0 so that one doesn't have to remember doing so.

Also centralize the `:port` setting so that tests which pass extra settings only have to declare the
ones which are relevant to the test.