Only point to :leak-detection Leiningen profile for enabling
Since this should be the overwhelmingly common way to enable the detector, we don't need to explain
how to enable it manually. Looking at the profile itself should be sufficient to understand which
settings are needed. Also, this way there's less of a chance for the profile and the docstring to
diverge over time.

Also, provide more context in the init log in case the logger doesn't print the current namespace by
default.