Support cancellation during client connection establishment
With #714 we added support for cancelling in-flight HTTP requests by putting the response deferred
into an error state. However, this only worked once the underlying TCP connection was
established. With this patch, it is now possible to cancel requests even while the connection is
still being established (possible since Netty 4.1.108.Final via
https://github.com/netty/netty/pull/13849). This also works for `aleph.tcp/client`.