Lift ssl-context coercion to connection-pool fn
As suggested by @DerGuteMoritz in #728. This fixes the issue and makes
the test added in the previous commit pass.

To avoid a repeated coercion *also* in `http-connection` fn (which is an
internal fn called only from `connection-pool` fn indirectly), the coercion
is being removed from it. As a result, we have to rely solely on
`connection-pool` fn for the coercion, where we don't know whether SSL
context will be required or not. This includes a situation when a user hasn't
supplied any `ssl-context` options, and we are still building an
`SslContext`, not knowing whether actual requests will require it or not.
This might look like a bit of potentially unnecessary work, but this
hopefully won't affect anyone performance-wise, as it would be unusual to run
`connection-pool` fn in a tight loop. On the positive side, all users will
profit from the elimination of repeated `SslContext` building, whether they
supply `ssl-context` options or not. Above all, having consistent code paths
in both situations is nice.

Notes:
- `http-connection` fn has now always some `ssl-context` available regardless
  of `ssl?` flag, but the only place where it matters
  (`client/make-pipeline-builder`) already checks `ssl?` flag before making
  use of `ssl-context`, so this makes no difference.
- All rich comments exercising `http-connection` fn with `ssl?` flag were
  updated, as now the caller has to supply `ssl-context` in that case.