Install client proxy handlers earlier
Since the introduction of HTTP/2 support in the client, proxy handlers (requested via proxy-options)
would be installed only after the HTTP version was negotiated. In case of a HTTPS connection, this
would be attempted via ALPN which means the TLS handshake would have already taken place with the
actual destination host, bypassing the configured proxy. This resulted in an error (see #731).

With this patch, the proxy handlers are installed earlier again so that a configured proxy is also
used during the TLS handshake phase.