Adopting an existing, pre-bound server-socket channel
The use-case is to pass a server-socket channel, e.g. obtained by
calling JVM's `System/inheritedChannel`. In this case, Aleph/Netty is
not responsible for opening the channel nor closing it. It should
simply listen on it. In other words, maintaining the connection
sockets is still Aleph/Netty's responsibility, but not maintaining
the listening socket if it has been passed to it.

* In the server startup part, this has been implemented as per Netty's
  Norman Maurer's advice [found here]
  (https://stackoverflow.com/a/50196956).

* In the server shutdown part, Aleph is not shutting down nor closing
  the passed socket, because in practice it's owned by a process that
  passed it to Aleph process, and it's that parent process'
  responsibility to reuse or close it after Aleph process exits.

In terms of validation and fallback, for comparison, [this used to be
Jetty's approach to adopting `inheritedChannel` at version 8]
(https://github.com/jetty/jetty.project/blob/jetty-8.0.0.v20110901/jetty-server/src/main/java/org/eclipse/jetty/server/nio/InheritedChannelConnector.java).

* Jetty automatically falls back to opening its own socket (with a
  warning) if the `inheritedChannel` is not supported or not present.
  In Aleph implementation, it refuses to start the server if the
  channel type is not supported, or the socket is not bound, but it
  silently falls back to an own socket if the passed channel is
  `nil`. As a result, it's Aleph user's responsibility to trigger an
  error if it expects an 'inherited channel' which is missing. Still,
  the user can expect that Aleph will fail early if a channel was
  actually passed, but it cannot be used.

* Unlike Jetty, Aleph doesn't itself call `configureBlocking false` on
  the channel because [this is done by Netty during server startup]
  (https://github.com/netty/netty/blob/46d6e164a2a46e2c6d802f84e0c24abd475bcd89/transport/src/main/java/io/netty/channel/nio/AbstractNioChannel.java#L84).