Don't leak unconsumed deferreds in error state in HTTP client
To that end, turn `wrap-exceptions` from a wrapping middleware into an "after handler" (like
`handle-redirects`) called `handle-error-status`.

This also happens to improve a suboptimal behavior: When a request with `:throw-exceptions? true`
was made and an error response was received which the middleware would turn into an exception, the
underlying connection would be closed regardless of whether the response indicated keep-alive or not.

Fixes #766 except for cases originating from Manifold's `d/timeout!` which needs to be fixed there.